+++ src/main/java/net/hecco/bountifulfares/block/custom/ChamomileCandleBlock.java
@@ -2,17 +2,19 @@ package net.hecco.bountifulfares.block.custom;
 
 import net.hecco.bountifulfares.block.entity.ChamomileCandleBlockEntity;
 import net.hecco.bountifulfares.block.entity.BFBlockEntities;
+import net.hecco.bountifulfares.effect.BFEffects;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public class ChamomileCandleBlock extends InfusedCandleBlock {
     public ChamomileCandleBlock(Settings settings) {
-        super(settings);
+        super(BFEffects.EBULLIENCE, settings);
     }
 
     @Nullable