+++ src/main/java/net/hecco/bountifulfares/block/custom/TorchflowerCandleBlock.java
@@ -6,13 +6,14 @@ import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public class TorchflowerCandleBlock extends InfusedCandleBlock {
     public TorchflowerCandleBlock(Settings settings) {
-        super(settings);
+        super(StatusEffects.STRENGTH, settings);
     }
 
     @Nullable