+++ src/main/java/net/hecco/bountifulfares/block/custom/WalnutCandleBlock.java
@@ -2,17 +2,19 @@ package net.hecco.bountifulfares.block.custom;
 
 import net.hecco.bountifulfares.block.entity.BFBlockEntities;
 import net.hecco.bountifulfares.block.entity.WalnutCandleBlockEntity;
+import net.hecco.bountifulfares.effect.BFEffects;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public class WalnutCandleBlock extends InfusedCandleBlock {
     public WalnutCandleBlock(Settings settings) {
-        super(settings);
+        super(BFEffects.ENRICHMENT, settings);
     }
 
     @Nullable