+++ src/main/java/net/hecco/bountifulfares/block/entity/GristmillBlockEntity.java
@@ -37,7 +37,7 @@ import java.util.Objects;
 import java.util.Optional;
 
 public class GristmillBlockEntity extends BlockEntity implements SidedInventory, ImplementedInventory, ExtendedScreenHandlerFactory<GristmillPayload>, RecipeInputProvider {
-
+    //TODO: fix gristmill recipes
     private static BooleanProperty millingState;
     private final DefaultedList<ItemStack> inventory = DefaultedList.ofSize(2, ItemStack.EMPTY);