+++ src/main/java/net/hecco/bountifulfares/item/custom/BlackTeaBottleItem.java
@@ -2,14 +2,16 @@ package net.hecco.bountifulfares.item.custom;
 
 import net.hecco.bountifulfares.effect.BFEffects;
 import net.minecraft.entity.effect.StatusEffect;
+import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.registry.entry.RegistryEntry;
 
 import java.util.ArrayList;
+import java.util.List;
 
 public class BlackTeaBottleItem extends TeaBottleItem{
     public BlackTeaBottleItem(Settings settings) {
-        super(settings);
+        super(List.of(new StatusEffectInstance(StatusEffects.WEAKNESS), new StatusEffectInstance(StatusEffects.SLOWNESS), new StatusEffectInstance(BFEffects.ACIDIC)), settings);
     }
     public ArrayList<RegistryEntry<StatusEffect>> getStatusEffectsToRemove() {
         ArrayList<RegistryEntry<StatusEffect>> list = new ArrayList<>();