+++ src/main/java/net/hecco/bountifulfares/item/custom/MeadBottleItem.java
@@ -1,15 +1,26 @@
 package net.hecco.bountifulfares.item.custom;
 
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.tooltip.TooltipType;
+import net.minecraft.screen.ScreenTexts;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
+import net.minecraft.text.Text;
+import net.minecraft.util.Formatting;
 import net.minecraft.world.World;
+
+import java.util.List;
 
 public class MeadBottleItem extends LiquidBottleItem {
     public MeadBottleItem(Settings settings) {
         super(settings);
+    }
+    public MeadBottleItem(List<StatusEffectInstance> effects, Settings settings) {
+        super(settings);
+        this.effects = effects;
     }
     public ItemStack finishUsing(ItemStack stack, World world, LivingEntity user) {
         super.finishUsing(stack, world, user);
@@ -22,5 +33,13 @@ public class MeadBottleItem extends LiquidBottleItem {
     @Override
     public SoundEvent getEatSound() {
         return SoundEvents.ENTITY_GENERIC_DRINK;
+    }
+
+    @Override
+    public void appendTooltip(ItemStack stack, TooltipContext context, List<Text> tooltip, TooltipType type) {
+        super.appendTooltip(stack, context, tooltip, type);
+        tooltip.add(ScreenTexts.EMPTY);
+        tooltip.add(Text.translatable("tooltip.bountifulfares.removes").formatted(Formatting.GRAY));
+        tooltip.add(Text.translatable(new StatusEffectInstance(StatusEffects.POISON).getTranslationKey().formatted(StatusEffects.POISON.value().getCategory().getFormatting())).formatted(Formatting.RED));
     }
 }